Move action for showing a detailed audit log page to the AuditController for better reuse
